:truck: Move theme models into theme store folder
