+++ shared/themes/dracula.ts
@@ -1,4 +1,4 @@
-import { ThemeColors } from '../models/themeColor.model';
+import { ThemeColors } from '../stores/theme/themeColor.model';
 
 const draculaTheme: ThemeColors = {
   light: '#FFF',
